requirements.txt
streamlit
ultralytics
pillow
pytube